// src/app/core/interfaces/iblog.ts
export interface IBlog {
  id: number;
  slug: string;
  featured_image: string;
  gallery: any;
  active: boolean;
  published_at: any;
  status: string;
  deleted_at: any;
  created_at: string;
  updated_at: string;
  display_order: number;
  title: string;
  description: string;
  tags: string;
}